+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/BlockGrpcClient.kt
@@ -1,28 +1,15 @@
 package io.provenance.explorer.grpc.v1
 
-import com.google.protobuf.util.JsonFormat
 import cosmos.base.tendermint.v1beta1.Query
 import cosmos.base.tendermint.v1beta1.ServiceGrpc
 import io.grpc.ManagedChannelBuilder
-import io.ktor.client.call.receive
-import io.ktor.client.features.ResponseException
-import io.ktor.client.request.get
-import io.ktor.client.statement.HttpResponse
-import io.provenance.explorer.KTOR_CLIENT_JAVA
-import io.provenance.explorer.config.ExplorerProperties
 import io.provenance.explorer.config.interceptor.GrpcLoggingInterceptor
-import io.provenance.explorer.domain.exceptions.FigmentApiException
-import kotlinx.coroutines.runBlocking
 import org.springframework.stereotype.Component
 import java.net.URI
 import java.util.concurrent.TimeUnit
 
 @Component
-class BlockGrpcClient(
-    channelUri: URI,
-    private val protoParser: JsonFormat.Parser,
-    private val explorerProps: ExplorerProperties
-) {
+class BlockGrpcClient(channelUri: URI) {
 
     private val tmClient: ServiceGrpc.ServiceBlockingStub
 
@@ -53,18 +40,4 @@ class BlockGrpcClient(
         }
 
     fun getLatestBlock() = tmClient.getLatestBlock(Query.GetLatestBlockRequest.getDefaultInstance())
-
-    fun getBlockAtHeightFromFigment(height: Int) = runBlocking {
-        val res = try {
-            KTOR_CLIENT_JAVA.get<HttpResponse>("${explorerProps.figmentUrl}/apikey/${explorerProps.figmentApikey}/cosmos/base/tendermint/v1beta1/blocks/$height")
-        } catch (e: ResponseException) {
-            throw FigmentApiException("Error reaching figment: ${e.response}")
-        }
-
-        if (res.status.value in 200..299) {
-            val builder = Query.GetBlockByHeightResponse.newBuilder()
-            protoParser.ignoringUnknownFields().merge(res.receive<String>(), builder)
-            builder.build()
-        } else throw FigmentApiException("Error reaching figment: ${res.status.value}")
-    }
 }